fix:rename the helper function
